Move recipients filtering in a new utils module
