+++ src/sib/transactional.py
@@ -1,21 +1,6 @@
 from anymail.message import AnymailMessage
 
-from django.conf import settings
-
-
-def filter_recipients(recipient_list):
-    """
-    Filter the recipients through a whitelist.
-    """
-    if not settings.ENABLE_EMAIL_WHITELIST:
-        return recipient_list
-    filtered_recipient = [
-        addr for addr in recipient_list
-        if addr in settings.EMAIL_WHITELIST
-    ]
-    print(filtered_recipient)
-    print(settings.EMAIL_WHITELIST)
-    return filtered_recipient
+from sib.utils import filter_recipients
 
 
 def send_mail(
